% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knownMotif.R
\name{annotKnownMotifs}
\alias{annotKnownMotifs}
\title{Annotate discovered motif with known motifs}
\usage{
annotKnownMotifs(x)
}
\arguments{
\item{x}{a \code{data.frame} has at least a column names as "motif"}
}
\value{
a \code{data.frame} has two other columns attached at the end 
  of original input \code{x} for known motifs and the corresponding kinases
}
\description{
Annotate discovered motif with known motifs
}
\references{
http://hprd.org/PhosphoMotif_finder
}
\author{
Chen Meng
}
