% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectMotifs.R
\name{detectMotifs}
\alias{detectMotifs}
\title{Detect over-represented motifs}
\usage{
detectMotifs(fg.seqs, bg.seqs, method = c("cvp", "all")[1], min.seqs = 5,
  max.fdr = 0.01, ncores = 1, verbose = TRUE)
}
\arguments{
\item{fg.seqs}{a character vector of foreground sequences}

\item{bg.seqs}{a character vector of background sequences}

\item{method}{the algorithm used to detect potential motifs, multiple are allowed}

\item{min.seqs}{the minimum frequency of a motif should be considered}

\item{max.fdr}{the maximum FDR to be reported}

\item{ncores}{the number of cores to be used, passed to \code{mclapply}.}
}
\value{
a \code{data.frame} of the over-representation data analysis
}
\description{
Detect over-represented motifs
}
\details{
more details here
}
\author{
Chen Meng
}
