% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_all.R
\name{motif_all}
\alias{motif_all}
\title{Find ALL motifs in a list of sequences using motif-all algorithm}
\usage{
motif_all(seqs, min.seqs, ncores = 1, verbose = FALSE)
}
\arguments{
\item{seqs}{a character vector of sequences}

\item{min.seqs}{minimum number of motif}

\item{ncores}{the number of cores to be used, passed to \code{mclapply}.}

\item{verbose}{logical, whether print detailed information}
}
\value{
a names integer vector. The names are the sequence motifs and
  the integers indicate the frequency of each motifs in the input sequences.
}
\description{
Find ALL motifs in a list of sequences using motif-all algorithm
}
\references{
He, Zengyou, Can Yang, Guangyu Guo, Ning Li, and Weichuan Yu. 2011.
  "Motif-All: Discovering All Phosphorylation Motifs." BMC Bioinformatics
  12 Suppl 1 (February):S22.
}
\author{
Chen Meng
}
