% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_exh.R
\name{motif_exh}
\alias{motif_exh}
\title{Find fixed position motifs (inconsecutive or consecutive) in a list of
  sequences using an exhausitive approach}
\usage{
motif_exh(seqs, min.seqs = 1, ncores = 1)
}
\arguments{
\item{seqs}{a character vector of sequences}

\item{min.seqs}{minimum number of motif}

\item{ncores}{the number of cores to be used, passed to \code{mclapply}.}
}
\value{
a names integer vector. The names are the sequence motifs and
  the integers indicate the frequency of each motifs in the input sequences.
}
\description{
Find fixed position motifs (inconsecutive or consecutive) in a list of
  sequences using an exhausitive approach
}
\note{
the algorithm uses an exhaustive approach to find all (long)
  fixed position motifs, so it could be slow when the number of sequences is large.
}
\author{
Chen Meng
}
