% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motifExampleData}
\alias{motifExampleData}
\title{Example dataset for motif analysis}
\format{A \code{list} with two elements:
\describe{
  \item{foreground}{foreground sequences}
  \item{background}{background sequences}
}}
\usage{
data(motifExampleData)
}
\description{
Example dataset for motif analysis
}
