% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifor.R
\name{motifgeneor}
\alias{motifgeneor}
\title{Find over-represented motifs in a foreground list using hypergeometric test}
\usage{
motifgeneor(motif, fg.genes, fg.genes.motif, bg.seqs, bg.genes)
}
\arguments{
\item{motif}{a motif}

\item{fg.genes}{unique foreground genes}

\item{fg.genes.motif}{unique foreground genes containing the motif}

\item{bg.seqs}{the background sequences}

\item{bg.genes}{a character vector of genes names of the bg.seqs, it
should have the same length as \code{bg.seqs}}
}
\description{
Find over-represented motifs in a foreground list using hypergeometric test
}
\author{
Chen Meng
}
