% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifor.R
\name{motifor}
\alias{motifor}
\title{Find over-represented motifs in a foreground list using hypergeometric test}
\usage{
motifor(fg.count, n.fg.seqs, bg.seqs, ncores = 1, max.fdr = 0.01)
}
\arguments{
\item{fg.count}{a names integer vector. The names are the sequence motifs and
the integers indicate the frequency of each motifs in the input sequences,
usually return by \code{\link{motif_exh}} or \code{\link{motif_cvp}}.}

\item{n.fg.seqs}{the number of foreground sequences}

\item{bg.seqs}{a character vector of background sequences}

\item{ncores}{the number of cores to be used, passed to \code{mclapply}.}

\item{max.fdr}{the maximum FDR to be reported}
}
\value{
a \code{data.frame} of the over-representation data analysis
}
\description{
Find over-represented motifs in a foreground list using hypergeometric test
}
\author{
Chen Meng
}
