% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSeqs.R
\name{subSeqs}
\alias{subSeqs}
\title{Find all subsequences of a sequence in give length(s)}
\usage{
subSeqs(x, k = 2:length(x), trim.char = NULL)
}
\arguments{
\item{x}{a character vector represents the sequence, each elment should be a single character}

\item{k}{the length of subsequences}

\item{trim.char}{the character to be excluded from the sequences}
}
\value{
a character vector contains all the subsequences
}
\description{
Find all subsequences of a sequence in give length(s)
}
\examples{
  subSeqs(x=LETTERS[1:5])
  subSeqs(x=sample(LETTERS, 400, replace = TRUE), k = c(2, 4))
}
\author{
Chen Meng
}
