% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSeqs.R
\name{checkSeqs}
\alias{checkSeqs}
\title{Internal function used to check the input foreground and background sequence vectors}
\usage{
checkSeqs(fg.seqs, bg.seqs, option = c("extend", "trim", "error")[1])
}
\arguments{
\item{fg.seqs}{a character vector of foreground sequences}

\item{bg.seqs}{a character vector of background sequences}

\item{option}{what should be done when some sequences in fg.seqs do not exist in the bg.seqs.
acceptable options are "extend" - extend the background list; "trim" - trim the foreground list,
i.e. remove the ones in foreground but not in background; "error" - stop and returns
error message.}
}
\value{
a list has two components: fg.seqs and bg.seqs
}
\description{
Internal function used to check the input foreground and background sequence vectors
}
\author{
Chen Meng
}
