% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectMotifs.R
\name{detectMotifs}
\alias{detectMotifs}
\title{Detect over-represented motifs}
\usage{
detectMotifs(fg.seqs, bg.seqs, fg.genes = NULL, bg.genes = NULL,
  method = c("cvp", "all")[1], min.seqs = 5, max.fdr = 0.01,
  center = "STY", ncores = 1, verbose = TRUE, annotate = FALSE)
}
\arguments{
\item{fg.seqs}{a character vector of foreground sequences}

\item{bg.seqs}{a character vector of background sequences}

\item{fg.genes}{character vector of gene names of foreground sequences, it
should have the same length as \code{fg.seqs}}

\item{bg.genes}{character vector of gene names of background sequences, it
should have the same length as \code{bg.seqs}}

\item{method}{the algorithm used to detect potential motifs, multiple are allowed}

\item{min.seqs}{the minimum frequency of a motif should be considered}

\item{max.fdr}{the maximum FDR to be reported}

\item{center}{the amino acid centered at the sequences, sequences with other center AA would be 
removed from the list. To disable this function, set \code{center = NULL}. Only used in 
\code{checkSeqs} and \code{motif_all}.}

\item{ncores}{the number of cores to be used, passed to \code{mclapply}.}

\item{verbose}{logical, whether print message}

\item{annotate}{whether the discovered motif should be annotated by known motifs}
}
\value{
a \code{data.frame} of the over-representation data analysis
}
\description{
Detect over-represented motifs
}
\details{
more details here
}
\author{
Chen Meng
}
