% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_cvp.R
\name{motif_cvp}
\alias{motif_cvp}
\title{Find consecutive but variable position (CVP) motif in a list of sequences}
\usage{
motif_cvp(seqs, min.seqs = 1, ncores = 1)
}
\arguments{
\item{seqs}{a character vector of sequences}

\item{min.seqs}{minimum number of motif}

\item{ncores}{the number of cores to be used, passed to \code{mclapply}.}
}
\value{
a names integer vector. The names are the sequence motifs and
  the integers indicate the frequency of each motifs in the input sequences.
}
\description{
Find consecutive but variable position (CVP) motif in a list of sequences
}
\note{
the algorithm uses an exhaustive approach to find all consecutive
  motifs, so it could be slow when the number of sequences is large.
}
\examples{
  seqs <- c(paste(LETTERS[1:6], collapse = ""),
  paste(LETTERS[2:7], collapse = ""),
  paste(LETTERS[3:8], collapse = ""),
  paste(LETTERS[4:9], collapse = ""),
  paste(LETTERS[5:10], collapse = ""))
  motif_cvp(seqs)
}
\author{
Chen Meng
}
